[refactor] Move eXist-db REST API External Variable Tests into their own class
